move the test patterns from utils.js to test-patterns.js
